refactor(web): correct dependency direction + decouple ArteOdyssey binding
Two follow-up findings from a deeper read of the previous refactor:

1. chat-pane.tsx (UI primitive) was importing ChatMessage from
   share-types.ts (server-side share concept). Conceptually inverted —
   chat existed before sharing did; share is a derivative use of chat.

   Fix: extract ChatMessage + chatMessageSchema to lib/chat-types.ts as
   the canonical chat primitive. share-types.ts now imports
   chatMessageSchema from chat-types and builds snapshotInputSchema on
   top of it. chat-pane / home-shell / share-button all import
   ChatMessage from chat-types directly, so the chat-pane re-export
   trick (which the previous PR introduced for compat) is gone.

2. arteOdysseyAdapter was hard-imported from
   '@decoro/adapter-arte-odyssey' in three places: /api/generate/route,
   code-panel, and /preview/page. Same anti-pattern that ADR-004 says
   to avoid — library-specific knowledge leaking into apps/web instead
   of staying behind the Adapter interface.

   Fix: pin the adapter binding in decoro.config.ts (same pattern as
   `llm`). Callers import `adapter` from the config and consume it
   through the abstract Adapter interface. The only place in apps/web
   that names @decoro/adapter-arte-odyssey is now the config itself —
   future @decoro/adapter-mui swap is a one-line change.

   Typed as `typeof arteOdysseyAdapter` (not the bare `Adapter`) so the
   registry's component type stays narrow — `Adapter` defaults
   `TComponent` to `unknown`, which json-render's `ComponentRegistry`
   would refuse.

Verified: typecheck / test (15/15) / lint (42 files, 0 issues) all
green. `grep` confirms one — and only one — reference to
@decoro/adapter-arte-odyssey across apps/web.
